# openapi.yml
openapi: '3.0.2'
info:
  title: LIU TIANQI blog system
  description: this is a blog system for LIU TIANQI
  version: '1.0'
servers:
  - url: https://localhost:8080
paths:
...
  /editor:
    get:
      tags:
        - blogs
      description: '返回editor界面'
      operationId: getEditView
      requestBody:
        required: true
        content:
          text/json:
            schema:
                type: "object"

                properties:
                  userId:
                    type: "string"
                  blogId:
                    type: "string"
      responses:
        '200':
          description: OK
          content:
            text/html:    
              schema:
                type: "string"
                example: "edit.html" 
    post:
      tags:
        - blogs
      description: '接受user输入的编辑信息,并返回blog页面'
      operationId: editor
      requestBody:
        required: true
        content:
          text/json:
            schema:
              type: "object"
              properties:
                title:
                  type: "string"
                introduction:
                  type: "string"
                contents:
                  type: "string"
      responses:
        '200':
          description: OK
          content:
            text/html:    
              schema:
               type: "object"

               properties:
                 username:
                  type: "string"
                 blogs:
                   type: array
                   items:
                      type: "object"

                      properties:
                        title:
                          type: "string"
                        introduction:
                          type: "string"
                        contents:
                          type: "string"
                 view:
                   type: "string"
                   example: "blog.html"
...